add docker to build / run the app moving stuff around added makefile
