+++ maestro-app/src/main/java/bio/overture/maestro/infra/adapter/inbound/webapi/ErrorDetails.java
