+++ maestro-app/src/main/java/bio/overture/maestro/infra/adapter/inbound/webapi/GlobalWebExceptionHandler.java
@@ -7,6 +7,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 import java.util.Date;
@@ -17,6 +18,7 @@ public class GlobalWebExceptionHandler {
 
     @ExceptionHandler(NotFoundException.class)
     @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ResponseBody
     public ErrorDetails resourceNotFoundException(NotFoundException ex, ServerHttpRequest request) {
         log.error("resource not found exception", ex);
         return getErrorDetails(ex, request);
@@ -24,6 +26,7 @@ public class GlobalWebExceptionHandler {
 
     @ExceptionHandler(Exception.class)
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ResponseBody
     public ErrorDetails globalExcpetionHandler(Exception ex, ServerHttpRequest request) {
         log.error("unhandled exception", ex);
         return getErrorDetails(ex, request);
