+++ maestro-app/src/main/resources/config/application.yml
@@ -3,8 +3,8 @@ server:
 
 maestro:
   elasticsearch:
-    cluster-name: maestro.elasticsearch
-    cluster-nodes: localhost:9200, localhost:9201
+    cluster-name: ${ELASTICSEARCH.CLUSTER.NAME:maestro.elasticsearch}
+    cluster-nodes: ${ELASTICSEARCH.CLUSTER.NODES:localhost:9200, localhost:9201}
     indexes:
       file-centric:
         alias: file-centric
@@ -21,3 +21,12 @@ maestro:
       url: http://aws.amazon.com
       name: aws repo
       country: US
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: '*'
+  endpoint:
+    health:
+    show-details: ALWAYS
