+++ maestro-app/src/main/resources/file-centric.mapping.json
