+++ maestro-app/src/main/resources/index.settings.json
