+++ maestro-app/src/main/resources/repository.mapping.json
