+++ maestro-app/src/test/resources/fixtures/ManagementControllerTest/study.json
