--- maestro-infra/docker/docker-compose.yml
@@ -1,11 +1,0 @@
-version: '3'
-services:
-  elasticsearch:
-    image: docker.elastic.co/elasticsearch/elasticsearch-oss:6.6.1
-    container_name: maestro.elasticsearch
-    ports:
-      - 9200:9200
-    environment:
-      - discovery.type=single-node
-      - cluster.name=maestro.elasticsearch
-      - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
